Refactor HTML attribute handling in CanvasNode and BoardPreviewNode
- Introduced a new utility function `getPropertyHtmlAttributes` to centralize the mapping of node properties to HTML attributes, improving code maintainability and readability.
- Updated `CanvasNode` to utilize the new utility for setting HTML attributes, reducing redundancy in attribute assignment.
- Modified `BoardPreviewNode` to incorporate the new utility, ensuring consistent attribute handling across components.

These changes streamline the process of rendering HTML attributes from node properties, enhancing the overall structure of the codebase.
