+++ packages/editor/app/canvas/Node.tsx
@@ -37,6 +37,7 @@ import {
 } from "@lib/workspace/node-tree"
 import { buildRenderParentIndex } from "@lib/workspace/render-parent-index"
 import { CanvasHtmlAttributes, ComponentRenderer } from "./ComponentRenderer"
+import { getPropertyHtmlAttributes } from "./property-html-attributes"
 import { REPEAT_OUTLINE } from "./canvas.bespoke"
 
 export type CanvasNodeProps = {
@@ -302,68 +303,7 @@ export const CanvasNode = memo(function CanvasNode({
       "data-selection-kind": "node",
       "data-selection-root-id": selfPath,
       "data-component-id": componentId ?? "",
-    }
-
-    // Some properties must be set as HTML attributes instead of tokens
-    if (properties.source?.value) {
-      htmlAttributes.src = properties.source.value
-    }
-
-    if (properties.altText?.value) {
-      htmlAttributes.alt = properties.altText.value
-    }
-
-    if (properties.placeholder?.value) {
-      htmlAttributes.placeholder = properties.placeholder.value
-    }
-
-    if (properties.ariaLabel?.value) {
-      htmlAttributes["aria-label"] = properties.ariaLabel.value
-    }
-
-    const ariaAttributeValues: Record<string, unknown> = {
-      role: properties.role?.value,
-      "aria-hidden": properties.ariaHidden?.value,
-      "aria-disabled": properties.ariaDisabled?.value,
-      "aria-expanded": properties.ariaExpanded?.value,
-      "aria-selected": properties.ariaSelected?.value,
-      "aria-checked": properties.ariaChecked?.value,
-      "aria-pressed": properties.ariaPressed?.value,
-      "aria-current": properties.ariaCurrent?.value,
-      "aria-haspopup": properties.ariaHasPopup?.value,
-      "aria-invalid": properties.ariaInvalid?.value,
-      "aria-required": properties.ariaRequired?.value,
-      "aria-readonly": properties.ariaReadonly?.value,
-      "aria-live": properties.ariaLive?.value,
-    }
-    for (const [attribute, value] of Object.entries(ariaAttributeValues)) {
-      if (value == null) continue
-      htmlAttributes[attribute] =
-        typeof value === "boolean" ? value : String(value)
-    }
-
-    if (properties.inputType?.value) {
-      htmlAttributes.type = properties.inputType.value
-    }
-
-    if (properties.checked?.value) {
-      htmlAttributes.defaultChecked = true
-    }
-
-    if (properties.columns?.value) {
-      const columnValue =
-        typeof properties.columns.value === "number"
-          ? properties.columns.value
-          : properties.columns.value.value
-      htmlAttributes.colSpan = columnValue.toString()
-    }
-
-    if (properties.rows?.value) {
-      const rowValue =
-        typeof properties.rows.value === "number"
-          ? properties.rows.value
-          : properties.rows.value.value
-      htmlAttributes.rowSpan = rowValue.toString()
+      ...getPropertyHtmlAttributes(properties),
     }
 
     return htmlAttributes
